Bug Fix - C63, C64, I15
Reworked how filtering occurs for the system in that it no longer does a
hard code on the filter values, but rather saves them to user data. Put
extra safe guards in the sql injection inputs so that it runs though
JRequest for all values to prevent the occurance of an injection.