Bug Fix - AJAX (Competitions) Removal / Update
Removed most AJAX from competitions and replaced it with a confirmation
system. It will make an extra step though gives the user a confirmation
message. Teams has been reworked so it correctly uses this system with
slight help from AJAX due to the nature of selecting a team.

Security update to controllers is still required, though once teams is
updated this can occur without any blockers.